UX: Add CPU overclock and display rate override settings
These settings can be used to improve framerate or set a new framerate
limit (different from the typical 30/60 FPS).

Many games work well with this, but using overrides will break games
that are not designed around FPS-independent physics. Therefore, submitting
compatibility reports is disabled if either of the overrides are active.

Co-authored-by: Hugo Locurcio <hugo.locurcio@hugo.pro>